pyproject.toml
[project]
name = "sesac-upstage-homework"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.5.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-serializer>=1.6.1",
    "uvicorn>=0.40.0",
]
